Move VariationSetBuilder to togglz-testing
Relates to #245
