+++ junit/src/main/java/org/togglz/junit/vary/VariationSet.java
@@ -1,25 +1,13 @@
 package org.togglz.junit.vary;
 
-import java.util.Set;
-
 import org.togglz.core.Feature;
 
 /**
- * This class represents a set of feature state variants. It is used to configure unit test that are executed with
- * {@link FeatureVariations}. The common implementation of this interface is {@link VariationSetBuilder} which allows to build
- * sets dynamically.
- * 
- * @author Christian Kaltepoth
- * 
- * @see FeatureVariations
- * @see VariationSetBuilder
+ * @see org.togglz.testing.vary.VariationSet
  * @param <F> The feature class
+ *
+ * @deprecated use {@link org.togglz.testing.vary.VariationSet} instead
  */
-public interface VariationSet<F extends Feature> {
-
-    /**
-     * Build the variant set data structure from the current configuration of the class.
-     */
-    Set<Set<F>> getVariants();
-
+@Deprecated
+public interface VariationSet<F extends Feature> extends org.togglz.testing.vary.VariationSet<F> {
 }
