+++ junit/src/main/java/org/togglz/junit/vary/Variations.java
@@ -4,6 +4,7 @@ import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
+import org.togglz.testing.vary.VariationSetBuilder;
 
 /**
  * This annotation is used if a test class is executed with {@link FeatureVariations}. The method annotated with this annotation
